[Step3] refactor : step2 리뷰 반영 - takeMore 의존성 제거
