+++ src/main/kotlin/blackjack/Main.kt
@@ -2,7 +2,7 @@ package blackjack
 
 import blackjack.domain.card.CardDeck
 import blackjack.domain.game.BlackJackGame
-import blackjack.domain.game.TakeMore
+import blackjack.view.TakeMore
 import blackjack.domain.player.PlayerMaker
 import blackjack.view.InputView
 import blackjack.view.ResultView
